refactor(prompt): move user prompts to UserPrompt class
