+++ src/main/kotlin/racingcar/Application.kt
@@ -1,16 +1,9 @@
 package racingcar
 
 fun main() {
-    println("Names of the cars (comma-separated):")
-    val cars = readln().split(",").map { run {
-        require(it.trim().length <= 5) { "Each name should be up to 5 characters"}
-        val car = Car(it.trim())
-        car
-    } }
-
-    println("How many rounds will be played?")
-    val rounds = readln().toIntOrNull()
-    require(rounds != null) {"The number of rounds should be a valid number"}
+    val prompt = UserPrompt()
+    val cars = prompt.promptCarNames()
+    val rounds = prompt.promptRounds()
 
     val race = Race(cars = cars, rounds = rounds)
